Automake, Make: devide the orignal Make parser into the two parsers
This is an applicaiton for pushLanguage/popLanguage.
