+++ Tmain/list-roles.d/stdout-expected.txt
@@ -2,8 +2,6 @@
 #
 #
 #
-Automake	I	included	included	on
-Automake	I	optional	optionally included	on
 Automake	d	program	directory for PROGRAMS primary	on
 Automake	d	man	directory for MANS primary	on
 Automake	d	ltlibrary	directory for LTLIBRARIES primary	on
@@ -37,8 +35,6 @@ Vera	h	local	local header	on
 #
 # all.*
 #
-Automake	I	included	included	on
-Automake	I	optional	optionally included	on
 Automake	d	program	directory for PROGRAMS primary	on
 Automake	d	man	directory for MANS primary	on
 Automake	d	ltlibrary	directory for LTLIBRARIES primary	on
