+++ Units/parser-automake.r/langstack.am.d/args.ctags
@@ -1,0 +1,4 @@
+--sort=no
+--language-force=Automake
+--fields=+rlK
+--extra=+r
