+++ win32/ctags_vs2013.vcxproj.filters
@@ -114,6 +114,9 @@
     <ClCompile Include="..\parsers\make.c">
       <Filter>Source Files\Parsers</Filter>
     </ClCompile>
+    <ClCompile Include="..\parsers\automake.c">
+      <Filter>Source Files\Parsers</Filter>
+    </ClCompile>
     <ClCompile Include="..\parsers\man.c">
       <Filter>Source Files\Parsers</Filter>
     </ClCompile>
@@ -505,6 +508,9 @@
       <Filter>Header Files</Filter>
     </ClInclude>
    <ClInclude Include="..\main\mybool.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+   <ClInclude Include="..\parsers\make.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
